refac(ffi): winscard: small refactoring
