+++ ffi/src/winscard/scard_handle.rs
@@ -1,14 +1,10 @@
-use std::iter::once;
 use std::mem::size_of;
 use std::slice::{from_raw_parts, from_raw_parts_mut};
 
 use ffi_types::winscard::{LpScardIoRequest, ScardContext, ScardHandle, ScardIoRequest};
-use ffi_types::{LpCVoid, LpDword, LpStr, LpWStr};
+use ffi_types::LpCVoid;
 use winscard::winscard::{IoRequest, Protocol, WinScard, WinScardContext};
 use winscard::{Error, ErrorKind, WinScardResult};
-
-use super::buff_alloc::copy_w_buff;
-use crate::winscard::buff_alloc::copy_buff;
 
 /// Represents smart card context handle.
 /// Additionally, it holds allocated buffers and created smart card handles.
@@ -158,34 +154,4 @@ pub unsafe fn copy_io_request_to_scard_io_request(
     pci_buffer.copy_from_slice(&io_request.pci_info);
 
     Ok(())
-}
-
-pub unsafe fn write_multistring_w(
-    context: &mut WinScardContextHandle,
-    readers: &[&str],
-    dest: LpWStr,
-    dest_len: LpDword,
-) -> WinScardResult<()> {
-    let buffer: Vec<u16> = readers
-        .iter()
-        .flat_map(|reader| reader.encode_utf16().chain(once(0)))
-        .chain(once(0))
-        .collect();
-
-    copy_w_buff(context, dest, dest_len, &buffer)
-}
-
-pub unsafe fn write_multistring_a(
-    context: &mut WinScardContextHandle,
-    readers: &[&str],
-    dest: LpStr,
-    dest_len: LpDword,
-) -> WinScardResult<()> {
-    let buffer: Vec<u8> = readers
-        .iter()
-        .flat_map(|reader| reader.as_bytes().iter().cloned().chain(once(0)))
-        .chain(once(0))
-        .collect();
-
-    copy_buff(context, dest, dest_len, &buffer)
 }
